mpp-backend/requirements.txt
django==3.0.8
django-rest-passwordreset==1.1.0
djangorestframework==3.11.1
psycopg2==2.8.5
drf-yasg==1.17.1
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.5.0
python-dotenv==0.14.0
django-extensions==3.0.9
celery==5.0.0
django-celery-beat==2.0.0
gunicorn==20.0.4
psycopg2-binary==2.8.5
sentry-sdk==0.19.1
pytest==6.1.2
pytest-django==4.1.0
pytest-cov==2.10.1
mixer==6.1.3
pytest-html==3.1.0
pytest-docker-compose==3.2.0
pytest-dotenv==0.5.2
PyJWT==1.7.1
pytest-rerunfailures==9.1.1
pandas==1.2.2